feat: moved auth client/key functions to auth.js
