+++ api/src/auth/auth.js
@@ -1,16 +1,20 @@
-const jwt = require("jsonwebtoken");
-const jwksRsa = require("jwks-rsa");
+const jwksClient = require("jwks-rsa");
+const { promisify } = require("util");
 
-const checkJwt = jwt({
-    secret: jwksRsa.expressJwtSecret({
-        cache: true,
-        rateLimit: true,
-        jwksRequestsPerMinute: 5,
-        jwksUri: `${process.env.AUTH0_DOMAIN}/.well-known/jwks.json`,
-    }),
-    audience: process.env.AUTH0_AUDIENCE,
-    issuer: process.env.AUTH0_DOMAIN,
-    algorithm: ["RS256"],
+const client = jwksClient({
+    jwksUri: `${process.env.AUTH0_DOMAIN}.well-known/jwks.json`,
 });
 
-module.exports = checkJwt;
+const incomingKey = promisify(client.getSigningKey);
+
+const getUtilKey = async (header) => {
+    const result = await incomingKey(header.kid);
+
+    return result.publicKey;
+};
+
+module.exports = {
+    client,
+    incomingKey,
+    getUtilKey,
+};
